Rename Hardhat base project config
